requirements.txt
matplotlib~=3.5.2
numpy~=1.23.1
pillow~=9.2.0
pytorch==1.12.1
scikit-learn~=1.1.1
tqdm~=4.64.0
beautifulsoup4~=4.11.1
scikit-image~=0.19.2
flake8==5.0.4
pytest==7.2.0
mypy==0.982
